Auto-refresh ESI for a newly added character
When a character is added via EVE SSO, immediately dispatch all the
refresh jobs for that character so its data is pulled right away, and
land the user on /characters/refresh to watch it populate.

- Extract the refresh dispatch (insert refresh_task rows + enqueue queue
  messages) into a shared dispatchRefresh helper.
- refreshEsi (the button) now calls it for all the user's characters.
- The /character/callback route calls it for the just-added character and
  redirects to the status page instead of /character.
